wip: load macro tests from strings
