--- packages/macro/test/fixtures/js-selectOrdinal/actual.js
@@ -1,8 +1,0 @@
-import { t, selectOrdinal } from '@lingui/macro'
-
-t`This is my ${selectOrdinal({
-  value: count,
-  one: "#st",
-  "two": `#nd`,
-  other: ("#rd")
-})} cat`
