--- packages/macro/test/fixtures/js-selectOrdinal/expected.js
@@ -1,5 +1,0 @@
-import i18n from "@lingui/core";
-
-i18n._("This is my {count, selectordinal, one {#st} two {#nd} other {#rd}} cat", {
-  count: count
-});
