+++ packages/macro/test/fixtures/js-t-continuation-character.expected.js
@@ -1,0 +1,3 @@
+import i18n from "@lingui/core";
+
+i18n._("Multiline with continuation");
