+++ packages/macro/test/fixtures/js-t-continuation-character.js
@@ -1,0 +1,4 @@
+import { t } from '@lingui/macro'
+
+t`Multiline\
+  with continuation`
