+++ packages/macro/test/fixtures/jsx-plural-select-nested.expected.js
