+++ packages/macro/test/fixtures/jsx-plural-select-nested.js
