--- packages/macro/test/fixtures/jsx-selectordinal/actual.js
@@ -1,18 +1,0 @@
-import { Trans, SelectOrdinal } from '@lingui/macro';
-
-<Trans>
-  This is my <SelectOrdinal
-    value={count}
-    one="#st"
-    two={`#nd`}
-    other={<strong>#rd</strong>}
-  /> cat.
-</Trans>;
-<Trans>
-  This is my <SelectOrdinal
-  value={user.numCats}
-  one="#st"
-  two={`#nd`}
-  other={<strong>#rd</strong>}
-/> cat.
-</Trans>;
