--- packages/macro/test/fixtures/jsx-trans-elements/actual.js
@@ -1,11 +1,0 @@
-import { Trans } from '@lingui/macro';
-
-<Trans>
-  Hello <strong>World!</strong><br />
-  <p>
-    My name is <a href="/about">{" "}
-    <em>{name}</em></a>
-  </p>
-</Trans>;
-<Trans>{<span>Component inside expression container</span>}</Trans>;
-<Trans>{<br />}</Trans>;
