--- packages/macro/test/fixtures/jsx-trans-production/actual.js
@@ -1,3 +1,0 @@
-import { Trans } from '@lingui/macro';
-
-<Trans id="msg.hello" defaults="Hello World" comment="Hello World" />
