--- packages/macro/test/fixtures/jsx-trans-production/expected.js
@@ -1,2 +1,0 @@
-import { Trans } from "@lingui/core";
-<Trans id="msg.hello" />;
