--- packages/macro/test/fixtures/jsx-trans-spread-child/actual.js
@@ -1,3 +1,0 @@
-import { Trans } from '@lingui/macro';
-
-<Trans>{...spread}</Trans>
