--- packages/macro/test/fixtures/jsx-trans-spread-child/expected.js
@@ -1,2 +1,0 @@
-import { Trans } from "@lingui/core";
-<Trans>{...spread}</Trans>;
