--- packages/macro/test/fixtures/jsx-trans-text/actual.js
@@ -1,7 +1,0 @@
-import { Trans } from '@lingui/macro';
-
-<span>Without translation</span>;
-<Trans>Hello World</Trans>;
-<Trans id="Hello World">Hello World</Trans>;
-<Trans id="msg.hello">Hello World</Trans>;
-<Trans id={msg} />;
