--- packages/macro/test/fixtures/jsx-trans-text/expected.js
@@ -1,6 +1,0 @@
-import { Trans } from "@lingui/core";
-<span>Without translation</span>;
-<Trans id="Hello World" />;
-<Trans id="Hello World" />;
-<Trans id="msg.hello" defaults="Hello World" />;
-<Trans id={msg} />;
