--- packages/macro/test/fixtures/jsx-whitespace/actual.js
@@ -1,18 +1,0 @@
-import { Trans } from "@lingui/macro";
-
-<Trans>
-  Strip whitespace around arguments: &quot;
-  {name}
-  &quot;
-</Trans>
-
-;<Trans>
-  Strip whitespace around tags, but keep{" "}
-  <strong>forced spaces</strong>
-  !
-</Trans>
-
-;<Trans>
-  Keep forced{"\n"}
-  newlines!
-</Trans>
