--- packages/macro/test/fixtures/jsx-whitespace/expected.js
@@ -1,8 +1,0 @@
-import { Trans } from "@lingui/core";
-<Trans id="Strip whitespace around arguments: &quot;{name}&quot;" values={{
-  name: name
-}} />;
-<Trans id="Strip whitespace around tags, but keep <0>forced spaces</0>!" components={{
-  0: <strong />
-}} />;
-<Trans id="Keep forced\n newlines!" />;
