--- packages/macro/test/fixtures/react/actual.js
@@ -1,9 +1,0 @@
-import { t, plural, Trans } from '@lingui/macro'
-
-const T = <Trans>Read <a href="/more" title={t`Full content of ${articleName}`}>more</a></Trans>
-
-const P = <a href="/about" title={plural({
-  value: count,
-  one: "# book",
-  other: "# books"
-})}>About</a>
