--- packages/macro/test/fixturesLegacy/jsx-number/actual.js
@@ -1,6 +1,0 @@
-import { Trans, DateFormat, NumberFormat } from '@lingui/macro';
-
-<Trans>The answer is <NumberFormat value={value} /></Trans>;
-<DateFormat value={value} />;
-<NumberFormat value={value} />;
-<NumberFormat value={value} format="percent" />;
