--- packages/macro/test/fixturesLegacy/jsx-number/expected.js
@@ -1,7 +1,0 @@
-import { Trans, DateFormat, NumberFormat } from "@lingui/react";
-<Trans id="The answer is {value,number}" values={{
-  value: value
-}} />;
-<DateFormat value={value} />;
-<NumberFormat value={value} />;
-<NumberFormat value={value} format="percent" />;
