--- packages/macro/test/fixturesLegacy/number-custom-format/expected.js
@@ -1,27 +1,0 @@
-(
-/*i18n*/
-{
-  id: "Format in variable {name,number,currency}",
-  formats: {
-    currency: currency
-  },
-  values: {
-    name: name
-  }
-});
-(
-/*i18n*/
-{
-  id: "One-off format {name,number,number0}, test format name collision {name,number,number1}",
-  formats: {
-    number0: {
-      digits: 4
-    },
-    number1: {
-      digits: 2
-    }
-  },
-  values: {
-    name: name
-  }
-});
