--- packages/macro/test/fixturesLegacy/number/actual.js
@@ -1,5 +1,0 @@
-import { t } from '@lingui/macro'
-
-t`The answer is ${t.number(name)}`;
-
-t`The interest rate is ${t.number(interest.rate, 'percent')}`;
