--- packages/macro/test/fixturesLegacy/number/expected.js
@@ -1,16 +1,0 @@
-(
-/*i18n*/
-{
-  id: "The answer is {name,number}",
-  values: {
-    name: name
-  }
-});
-(
-/*i18n*/
-{
-  id: "The interest rate is {0,number,percent}",
-  values: {
-    0: interest.rate
-  }
-});
